Add files via upload
Given the abundance of media channels competing for the attention of users today, it would be useful to find out which channels are able to attract more consumers than others. Also, understanding the profile of customers consuming different media channels could help target customer segments of interest. The purpose of this project is to analyze a sample network containing media (newspaper and TV) channels and users of these channels.